More Small fixes for hinting

+++ src/cosim_toolbox/cosim_toolbox/helicsConfig.py
@@ -4,7 +4,7 @@ import json
 
 class HelicsMsg(object):
 
-    def __init__(self, name: str, period: int):
+    def __init__(self, name: str, period: float):
         # change logging to debug, warning, error
         self._subs = []
         self._pubs = []
